l46-l47: FAT16 + disk in the build (Lectures 46 + 47)
PeachOS64BitCourse commits 5f9ceeb (L46) and 494e092 (L47)
ported together. Upstream's L46 alone is unbuildable - it
adds fs/* to FILES but the fat16 / file C code calls
diskstreamer_* / disk_get which live in disk/, added in L47.
SamOs combines both so every commit on module1-64bit is
buildable.

Makefile: fs/file.o, fs/pparser.o, fs/fat/fat16.o,
disk/disk.o, disk/streamer.o added to FILES with build rules.

build.sh: mkdir build/fs build/fs/fat build/disk.

src/kernel.h:
  ISERR widened to cast through int64_t before truncating to
  int:
    OLD: #define ISERR(value) ((int)value < 0)
    NEW: #define ISERR(value) ((int)((int64_t)(value)) < 0)
  Necessary because the bare (int)value truncates the high 32
  bits, and an ERROR(-EIO)-style round trip lands a negative
  int64_t with the sign bit in the high half. Truncating drops
  the sign and ISERR returns false on a real error.

No source-file edits in fs/ or disk/. The 32-bit code is
pointer-width-clean (uses uintptr_t / uint32_t consistently;
no (unsigned int)ptr truncations) so it recompiles under
x86_64-elf-gcc as-is. ISERR was the only round-trip-correctness
fix needed at this layer.

Nothing in kernel_main actually opens a file or reads a
sector yet. Test asserts the same tokens as L38..L45 - the
unit of test is "the link succeeds".

Verified on VGA:
  Hello 64-bit!
  e820 total: 267910144
  ABCmultiheap ready
  hello
  Divide by zero error

Doc: docs/64bit/chapters/L46-L47-fs-disk-linked.md walks the
ISERR widening rationale and notes that no source edits were
needed in fs/ or disk/.

Next: Lecture 48 - process functionality into 64-bit.

+++ src/kernel.h
@@ -11,7 +11,11 @@
 // 64-bit pointer width without sign-extending nonsense.
 #define ERROR(value)    (void*)((intptr_t)(value))
 #define ERROR_I(value)  (int)((intptr_t)(value))
-#define ISERR(value)    ((int)value < 0)
+// L46 - cast through int64_t before truncating to int so that
+// a 64-bit error code with the high bit set survives the
+// negativity check intact. Necessary for ERROR(value) round
+// trips when the kernel grows >2 GiB heap pointers.
+#define ISERR(value)    ((int)((int64_t)(value)) < 0)
 
 void kernel_main();
 void print(const char* str);
